# Conflicts:
#	be/src/main/java/yeonba/be/mypage/dto/response/UserProfileDetailResponse.java
#	be/src/main/java/yeonba/be/user/dto/response/UserProfileResponse.java
#	be/src/main/java/yeonba/be/user/entity/User.java

+++ be/src/main/java/yeonba/be/user/dto/response/UserProfileResponse.java
@@ -54,7 +54,7 @@ public class UserProfileResponse {
       type = "number",
       description = "사진 싱크로율",
       example = "80")
-  private double photoSyncRate;
+  private int photoSyncRate;
 
   @Schema(
       type = "string",
